+++ src/config/cproutes/common.php
@@ -76,4 +76,5 @@ return [
     'plugin-store' => 'plugin-store',
     'plugin-store/callback' => 'plugin-store/callback',
     'plugin-store/<url:(.*)>' => 'plugin-store',
+    'components/preview/<templatePath:(.*)>' => 'components/preview',
 ];