+++ src/web/assets/cp/src/js/UI.js
@@ -1,52 +1,11 @@
 /** global: Craft */
 /** global: Garnish */
-Craft.ui = {
-  createButton: function (config) {
-    const $btn = $('<button/>', {
-      type: config.type || 'button',
-      class: 'btn',
-    });
-    if (config.id) {
-      $btn.attr('id', config.id);
-    }
-    if (config.class) {
-      $btn.addClass(config.class);
-    }
-    if (config.ariaLabel) {
-      $btn.attr('aria-label', config.ariaLabel);
-    }
-    if (config.role) {
-      $btn.attr('role', config.role);
-    }
-    if (config.html) {
-      $btn.html(config.html);
-    } else if (config.label) {
-      $btn.append($('<div class="label"/>').text(config.label));
-    } else {
-      $btn.addClass('btn-empty');
-    }
-    if (config.toggle) {
-      $btn.attr('aria-expanded', 'false');
-    }
-    if (config.controls) {
-      $btn.attr('aria-controls', config.controls);
-    }
-    if (config.spinner) {
-      $btn.append($('<div class="spinner spinner-absolute"/>'));
-    }
-    return $btn;
-  },
 
-  createSubmitButton: function (config) {
-    const $btn = this.createButton(
-      Object.assign({}, config, {
-        type: 'submit',
-        label: config.label || Craft.t('app', 'Submit'),
-      })
-    );
-    $btn.addClass('submit');
-    return $btn;
-  },
+import {createButton, createSubmitButton} from './ui/buttons';
+
+Craft.ui = {
+  createButton,
+  createSubmitButton,
 
   createTextInput: function (config) {
     config = $.extend(